backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.0.3

# System & Resource Monitoring (Optional)
# psutil==5.9.6  # Uncomment for worker resource monitoring

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0

# Logging & Monitoring
structlog==23.2.0

# Task Queue & Background Processing
celery==5.3.4  # Optional: if using Celery for some tasks
redis==5.0.1   # Optional: for caching and pub/sub

# Production
gunicorn==21.2.0